requirements.txt
llama_index.experimental
llama_index
openai
pandas
python-dotenv